Effeciency: Re-order collective_creation in play
By moving it before the creation role, it allows the subject_host role
to be only be specified once.  Though another play is still needed for
accessability checking and dependency installation, since the
runtime inventory may not be in-sync with the dynamic inventory.

Also, initialize the join_groups list in invcache to the empty-list
instead of None.  This makes for a better debug message.

Signed-off-by: Chris Evich <cevich@redhat.com>